Change how the API deals with error situations.
This changes the API to always allow adding/removing geofences, even when
no permission was given. Geofences just won't be monitored unless and
until the user gives permission to do so.

Also clarify some security and privacy considerations, and attempt to
improve the description of GeofenceErrorEvent to address some of what was
mentioned in issue #23.